Trait dslab_models::power::cpu::CpuPowerModel

pub trait CpuPowerModel: DynClone {
    // Required method
    fn get_power(
        &self,
        utilization: f64,
        frequency: Option<f64>,
        state: Option<usize>
    ) -> f64;
}

A model for estimating the power consumption of CPU based on its utilization and, optionally, current CPU frequency and state. Can also be used to model the power consumption of a whole host based on CPU utilization.

Required Methods

fn get_power( &self, utilization: f64, frequency: Option<f64>, state: Option<usize> ) -> f64

Returns CPU power consumption in Watts.

CPU utilization is passed as a float in [0, 1] range (1 - maximum utilization).

(Optional) Relative CPU frequency is passed as a float in [0, 1] range, where 0 corresponds to the minimum CPU frequency and 1 corresponds to the maximum CPU frequency.

(Optional) CPU power management state is passed as a numerical index in [0, num states) range.