Struct dslab_models::power::cpu_models::state_based::StateBasedCpuPowerModel

pub struct StateBasedCpuPowerModel { /* private fields */ }

A power model that takes into account CPU power management states by using a separate sub-model for each state.

Most modern processors have power management states such as P-states and C-states for Intel® processors. P-states are states (voltage-frequency pairs) that are used when CPU is busy to reduce its power by slowing it down and reducing its voltage. C-states are idle power saving states that shutdown the parts of CPU when it is unused.

This abstract model allows to specify an arbitrary set of CPU states and associate an arbitrary power model with each of these states. When the model is invoked via get_power() method, it obtains the current CPU state (passed as an index in [0, num states) range) from the state argument and invokes the corresponding model.

Implementations

impl StateBasedCpuPowerModel

pub fn new(state_models: Vec<Box<dyn CpuPowerModel>>) -> Self

Creates a state-based power model.

state_models - Vector holding a power model for each modeled state (state is referred by its index).