Trait dslab_models::power::memory::MemoryPowerModel

pub trait MemoryPowerModel: DynClone {
    // Required methods
    fn get_power_simple(&self, utilization: f64) -> f64;
    fn get_power_advanced(&self, read_util: f64, write_util: f64) -> f64;
}

A model for estimating the power consumption of memory based on its utilization.

Required Methods

fn get_power_simple(&self, utilization: f64) -> f64

Returns memory power consumption in Watts based on a single utilization value.

Memory utilization should be passed as a float in 0.0-1.0 range.

fn get_power_advanced(&self, read_util: f64, write_util: f64) -> f64

Returns memory power consumption in Watts based on separate read and write utilization values.

Memory read and write utilization should be passed as floats in 0.0-1.0 range. Has more priority than get_power_simple().