Struct dslab_models::throughput_sharing::FairThroughputSharingModelWithCancel

pub struct FairThroughputSharingModelWithCancel<T> { /* private fields */ }

Fast implementation of fair throughput sharing model with ability to cancel activities.

According to our tests, it is slightly (10-20%) slower than FairThroughputSharingModel.

Implementations

impl<T> FairThroughputSharingModelWithCancel<T>

pub fn new( throughput_function: ResourceThroughputFn, factor_function: Box<dyn ActivityFactorFn<T>>, ) -> Self

Creates model with given throughput and factor functions.

pub fn with_fixed_throughput(throughput: f64) -> Self

Creates model with fixed throughput.

pub fn with_dynamic_throughput( throughput_function: ResourceThroughputFn, ) -> Self

Creates model with dynamic throughput, represented by given closure.