Trait dslab_models::throughput_sharing::ThroughputSharingModel

pub trait ThroughputSharingModel<T> {
    // Required methods
    fn insert(
        &mut self,
        item: T,
        volume: f64,
        ctx: &SimulationContext,
    ) -> ActivityId;
    fn pop(&mut self) -> Option<(f64, T)>;
    fn peek(&mut self) -> Option<(f64, &T)>;

    // Provided method
    fn cancel(
        &mut self,
        _id: ActivityId,
        _ctx: &SimulationContext,
    ) -> Option<(f64, T)> { ... }
}

Trait for throughput sharing model.

Required Methods

fn insert( &mut self, item: T, volume: f64, ctx: &SimulationContext, ) -> ActivityId

Adds new activity into the model.

Activity is represented by item, has amount of work volume and starts at ctx.time().

fn pop(&mut self) -> Option<(f64, T)>

Returns the next activity completion time (if any) along with corresponding activity item.

The returned activity is removed from the model.

fn peek(&mut self) -> Option<(f64, &T)>

Returns the next activity completion time (if any) along with corresponding activity item.

In contrast to pop, the returned activity is not removed from the model.

Provided Methods

fn cancel( &mut self, _id: ActivityId, _ctx: &SimulationContext, ) -> Option<(f64, T)>

Cancels the activity with given id at ctx.time().

If the activity is still running returns the activity and the completed work volume, otherwise returns None.