Trait dslab_models::throughput_sharing::ThroughputSharingModel

pub trait ThroughputSharingModel<T> {
    // Required methods
    fn insert(&mut self, item: T, volume: f64, ctx: &mut SimulationContext);
    fn pop(&mut self) -> Option<(f64, T)>;
    fn peek(&self) -> Option<(f64, &T)>;
}

Trait for throughput sharing model.

Required Methods

fn insert(&mut self, item: T, volume: f64, ctx: &mut SimulationContext)

Adds new activity into the model.

Activity is represented by item, has amount of work volume and starts at ctx.time().

fn pop(&mut self) -> Option<(f64, T)>

Returns the next activity completion time (if any) along with corresponding activity item.

The returned activity is removed from the model.

fn peek(&self) -> Option<(f64, &T)>

Returns the next activity completion time (if any) along with corresponding activity item.

In contrast to pop, the returned activity is not removed from the model.